replaced some of the uint8 to int conversions which I actually needed

+++ Machine_Learning/Training_Framework.py
@@ -653,9 +653,9 @@ class Training_Framework():
     def FromTorchTraining(self, image):
         #Returns a trainable tensor back into a visual image.
         if self.Settings["Do norm"]:
-            return self.Reverse_Normalization(image).permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.int).numpy()
+            return self.Reverse_Normalization(image).permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.uint8).numpy()
         else:
-            return image.permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.int).numpy()
+            return image.permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.uint8).numpy()
 
 
 
@@ -811,9 +811,9 @@ class Model_Inference():
     def FromTorchTraining(self, image):
         #Returns a trainable tensor back into a visual image.
         if self.Settings["Do norm"]:
-            return self.Reverse_Normalization(image).permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.int).numpy()
+            return self.Reverse_Normalization(image).permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.uint8).numpy()
         else:
-            return image.permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.int).numpy()
+            return image.permute(1,2,0).mul_(255).clamp(0,255).to("cpu", torch.uint8).numpy()
         
     def Inference_run(self, runs=30):
         """
